+++ R/env.R
@@ -40,6 +40,13 @@ set_envvar <- function(envs, action = "replace") {
 #' @details if `NA` is used those environment variables will be unset.
 #' If there are any duplicated variable names only the last one is used.
 #' @seealso [Sys.setenv()]
+#' @examples
+#' with_envvar(new = c("GITHUB_PAT" = "abcdef"), Sys.getenv("GITHUB_PAT"))
+#'
+#' # with_envvar unsets variables after usage
+#' Sys.getenv("TEMP_SECRET")
+#' with_envvar(new = c("TEMP_SECRET" = "secret"), Sys.getenv("TEMP_SECRET"))
+#' Sys.getenv("TEMP_SECRET")
 #' @export
 with_envvar <- with_(set_envvar)