+++ tests/testthat/test-defer.R
@@ -36,11 +36,11 @@ test_that("defer()'s global env facilities work", {
   expect_length(h, 2)
   expect_equal(Sys.getenv("abcdefg"), "tuvwxyz")
 
-  expect_output(run_global_deferred(), "howdy")
+  expect_output(deferred_run(globalenv()), "howdy")
   expect_equal(Sys.getenv("abcdefg"), "abcdefg")
 
   defer(print("never going to happen"), envir = globalenv())
-  clear_global_deferred()
+  deferred_clear(globalenv())
 
   h <- get_handlers(globalenv())
   expect_null(h)